requirements.txt
numpy
scipy
tensorboard
librosa==0.9.2
numba==0.56.4
pytorch-lightning
gradio==3.38.0
gradio_client==0.8.1
ffmpeg-python
onnxruntime
tqdm
funasr==1.0.0
cn2an
pypinyin
pyopenjtalk
g2p_en
torchaudio
modelscope==1.10.0
sentencepiece
transformers
chardet
PyYAML
psutil
jieba_fast
jieba
LangSegment>=0.2.0
Faster_Whisper
wordsegment
sentence_transformers
pyloudnorm
openai-whisper